refactor forward logic + add websocket forwarding app logic
